+++ src/metadata/server/index.ts
@@ -5,6 +5,8 @@ export * from './jwt';
 export * from './orm';
 export * from './provide';
 export * from './router';
-export * from './server';
 export * from './ssl';
+export * from './views';
+
+export * from './server';
 export * from './task';